+++ R/tar_manifest.R
@@ -13,7 +13,7 @@
 #'   [all_of()] and [starts_with()].
 #' @param fields Names of the fields, or columns, to show. Set to `NULL` to
 #'   show all the fields (default). Otherwise, you can supply
-#'   symbols, a character vector, or `tidyselect` helpers like [starts_with()].
+#'   `tidyselect` helpers like [starts_with()].
 #'   Set to `NULL` to print all the fields.
 #'   The name of the target is always included as the first column
 #'   regardless of the selection.
@@ -46,6 +46,8 @@
 #'   * `cue_iteration`: Iteration cue from [tar_cue()].
 #'   * `packages`: List columns of packages loaded before building the target.
 #'   * `library`: List column of library paths to load the packages.
+#' @param drop_missing Logical of length 1, whether to automatically omit
+#'   empty columns and columns with all missing values.
 #' @examples
 #' if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) {
 #' tar_dir({ # tar_dir() runs code from a temporary directory.
@@ -67,18 +69,23 @@
 #' }
 tar_manifest <- function(
   names = NULL,
-  fields = c("name", "command", "pattern"),
+  fields = tidyselect::all_of(c("name", "command", "pattern")),
+  drop_missing = TRUE,
   callr_function = callr::r,
   callr_arguments = targets::tar_callr_args_default(callr_function),
   envir = parent.frame(),
   script = targets::tar_config_get("script")
 ) {
   force(envir)
+  tar_assert_lgl(drop_missing)
+  tar_assert_scalar(drop_missing)
+  tar_assert_none_na(drop_missing)
   tar_assert_callr_function(callr_function)
   tar_assert_list(callr_arguments, "callr_arguments mut be a list.")
   targets_arguments <- list(
     names_quosure = rlang::enquo(names),
-    fields_quosure = rlang::enquo(fields)
+    fields_quosure = rlang::enquo(fields),
+    drop_missing = drop_missing
   )
   callr_outer(
     targets_function = tar_manifest_inner,
@@ -95,7 +102,8 @@ tar_manifest <- function(
 tar_manifest_inner <- function(
   pipeline,
   names_quosure,
-  fields_quosure
+  fields_quosure,
+  drop_missing
 ) {
   igraph <- pipeline_produce_igraph(pipeline, targets_only = TRUE)
   all_names <- topo_sort_igraph(igraph)
@@ -105,7 +113,15 @@ tar_manifest_inner <- function(
   out <- do.call(rbind, out)
   fields <- tar_tidyselect_eval(fields_quosure, colnames(out)) %|||%
     colnames(out)
-  out[, base::union("name", fields), drop = FALSE]
+  out <- out[, base::union("name", fields), drop = FALSE]
+  if (drop_missing) {
+    for (field in colnames(out)) {
+      if (all(is.na(unlist(out[[field]])))) {
+        out[[field]] <- NULL
+      }
+    }
+  }
+  out
 }
 
 tar_manifest_target <- function(target) {
@@ -137,7 +153,7 @@ tar_manifest_target <- function(target) {
 }
 
 tar_manifest_command <- function(expr) {
-  out <- tar_deparse_safe(expr, collapse = " \\n ")
+  out <- tar_deparse_safe(expr, collapse = "\n ")
   out <- mask_pointers(out)
   string_sub_expression(out)
 }