+++ NEWS.md
@@ -3,6 +3,8 @@
 ## Enhancements
 
 * Simplify and improve error handling.
+* In the `command` column of `tar_manifest()` output, separate lines with "\n" instead of "\\n" so the text output is straightforward to work with.
+* Add a `drop_missing` argument to `tar_manifest()` to hide/show columns with all `NA` values.
 
 # targets 0.13.1